+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 **Breaking changes:**
 - Dropped support for Rails 4.2
 
+**Enhancements:**
+- Komponent now reports component stats when you run `bin/rails stats`
+
 **Bug fixes:**
 - Removed redundant `class` attribute in HAML templates